+++ packages/core/cypress/integration/analytics.test.js
@@ -263,7 +263,7 @@ describe('view_item_list event', () => {
     cy.visit(pages.collection, options)
     cy.waitForHydration()
 
-    cy.get('.product-grid [data-testid=product-link]').then(() => {
+    cy.get('[data-fs-product-grid] [data-testid=product-link]').then(() => {
       cy.scrollTo('bottom', { duration: 500 }).then(() => {
         cy.scrollTo('top', { duration: 500 }).then(() => {
           dataLayerHasEvent('view_item_list')