+++ packages/core/src/components/product/ProductCard/ProductCard.tsx
@@ -15,6 +15,7 @@ import Price from 'src/components/ui/Price'
 import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
 import { useProductLink } from 'src/sdk/product/useProductLink'
 import type { ProductSummary_ProductFragment } from '@generated/graphql'
+import styles from 'src/components/product/ProductCard/product-card.module.scss'
 
 type Variant = 'wide' | 'default'
 
@@ -56,6 +57,7 @@ function ProductCard({
       data-fs-product-card-bordered={bordered}
       data-fs-product-card-actionable={!!ButtonBuy}
       data-fs-product-card-sku={sku}
+      className={styles.fsProductCard}
       {...otherProps}
     >
       <UIProductCardImage>