+++ packages/core/src/components/skeletons/ProductGridSkeleton/ProductGridSkeleton.tsx
@@ -1,5 +1,6 @@
 import type { PropsWithChildren } from 'react'
 
+import styles from 'src/components/product/ProductGrid/product-grid.module.scss'
 import { ITEMS_PER_PAGE } from 'src/constants'
 
 import ProductCardSkeleton from '../ProductCardSkeleton'
@@ -13,7 +14,7 @@ function ProductGridSkeleton({
   loading = true,
 }: PropsWithChildren<Props>) {
   return loading ? (
-    <ul className="product-grid">
+    <ul className={styles.fsProductGrid}>
       {Array.from({ length: ITEMS_PER_PAGE }, (_, index) => (
         <li key={String(index)}>
           <ProductCardSkeleton bordered />